% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_AuxDataTransfer.R
\name{ApplyFilterBuildingDatasets}
\alias{ApplyFilterBuildingDatasets}
\title{Filter the building data tables by a list of dataset IDs}
\usage{
ApplyFilterBuildingDatasets(
  Data_Input,
  Data_Output_PreCalculated,
  Data_Output,
  Data_Calc,
  Header_BuildingData,
  DF_FilterBuildingData,
  myFilterName = "All"
)
}
\arguments{
\item{Data_Input}{the 1st building dataframe to be filtered}

\item{Data_Output}{the 3rd building dataframe to be filtered}

\item{Data_Calc}{the 4th building dataframe to be filtered}

\item{Header_BuildingData}{a data frame with meta information (no changes will be appled)}

\item{DF_FilterBuildingData}{a data frame containing the available filter lists}

\item{myFilterName}{a character string specifying the filter list to be used}

\item{Data_Output_Precalculated}{the 2nd building dataframe to be filtered}
}
\value{
myBuildingDataTables     a list of the dataframes used as parameter of the function,
                                 4 tables modified if a filter is applied
}
\description{
This function provides a filtered version of all 4 building MOBASY data tables
}
\examples{

myBuildingDataTables <-
ApplyFilterBuildingDatasets (
  Data_Input                = myBuildingDataTables$Data_Input,
  Data_Output_PreCalculated = myBuildingDataTables$Data_Output_PreCalculated,
  Data_Output               = myBuildingDataTables$Data_Output,
  Data_Calc                 = myBuildingDataTables$Data_Calc,
  Header_BuildingData       = myBuildingDataTables$Header_BuildingData,
  DF_FilterBuildingData     = myBuildingDataTables$DF_FilterBuildingData,
  myFilterName              = "MOBASY-Sample"
)


}
