% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Main.R
\name{EnergyProfileCalc}
\alias{EnergyProfileCalc}
\title{Perform Energy Profile calculation - use state indicators to estimate the input data of a physical model and calculate the energy demand for heating and DHW}
\usage{
EnergyProfileCalc(
  TabulaTables,
  myBuildingDataTables,
  Indicator_Include_UncertaintyAssessment = 0
)
}
\arguments{
\item{TabulaTables}{a list of dataframes with parameters used for the calculation}

\item{myBuildingDataTables}{a list of dataframes including the calculation input data
"Data_Input", an empty dataframe "Data_Output" providing the structure for the output,
the dataframe "Data_Output_PreCalculated" providing data calculated by the Excel tool
(useful for comparison by developers) and the dataframe "Data_Calc" which is used to
collect all variables and their values used in the different calculation functions.}
}
\value{
myOutputTables a list of two output dataframes Data_Output (main results in
a predefined structure) and Data_Calc (values of all temporary variables)
}
\description{
EnergyProfileCalc consists of a physical model for calculating the energy demand for heating and domestic hot water.
Energy Profile Indicators or detailed building data may be used as input data.
The energy performance is calculated by use of a standard climate.
NOT included is the use of local climate by postcode, the calculation of uncertainties
and a comparison with measured consumption data (target/actual comparison).
}
\details{
An overview of the method can be found in:
Loga, Tobias; Stein, Britta; Behem, Guillaume: Use of Energy Profile Indicators to Determine
the Expected Range of Heating Energy Consumption; Proceedings of the Conference
"Central Europe towards Sustainable Building" 2022 (CESB22),
4 to 6 July 2022; Acta Polytechnica CTU Proceedings 38:470–477, 2022
https://ojs.cvut.cz/ojs/index.php/APP/article/view/8299/6839
https://doi.org/10.14311/APP.2022.38.0470
}
\examples{

## Get local parameter tables
TabulaTables <- GetParameterTables_LocalExcel ()


## Get local building data
# Different options of dataset selection:

# (1) Get all datasets from the MOBASY sample
myBuildingDataTables <- GetBuildingData_LocalExcel ("MOBASY-Sample")

# (2) Get dataset of 1 building (example for webtool)
myBuildingDataTables <- GetBuildingData_LocalExcel ("WebTool")


## Calculate energy performance (Energy Profile procedure)

myOutputTables <- EnergyProfileCalc (
   TabulaTables,
   myBuildingDataTables
   )


## Show structure and content of the two output dataframes

str (myOutputTables$Data_Output)
str (myOutputTables$Data_Calc)

}
