% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_AuxDataTransfer.R
\name{GetBuildingData_RDataPackage}
\alias{GetBuildingData_RDataPackage}
\title{Get MOBASY building data from local Excel file}
\usage{
GetBuildingData_RDataPackage(myFilterName = "All")
}
\arguments{
\item{myFilterName}{a character string using predefined lists of datasets.
The currently defined subsets are:
"Examples":         Several example datasets
"MOBASY-Sample":    Datasets of more than 100 multi-family houses used for target / actual
                    comparison and benchmarking in the MOBASY project
                    (see https://www.iwu.de/forschung/energie/mobasy/)
"MOBASY-Sample_Checked-OnSite":    Datasets of 12 buildings from the MOBASY sample which were checked onsite
                    Can be used to test the function MobasyCalc () including target actual comparison
"ParameterStudy-CESB-2022":  10 datasets of the parameter study, published in
                    Loga, Tobias; Stein, Britta; Behem, Guillaume (2023):
                    Use of Energy Profile Indicators to Determine the Expected Range
                    of Heating Energy Consumption;
                    Proceedings of the Conference "Central Europe towards Sustainable Building"
                    2022 (CESB22), 4 to 6 July 2022;
                    Acta Polytechnica CTU Proceedings 38:470–477, 2022, published 2023
                    https://doi.org/10.14311/APP.2022.38.0470
                    (also implemented in the MOBASY project)
"ParameterStudy-PHSP-2023": 32 datasets of a parameter study on the
                   "PassivHausSozialPlus" buildings, implemented in the MOBASY project
"MOBASY-All"        All datasets mentioned above
"WebTool":          One dataset used for the webtool
"Typology-DE_Example-Buildings"  Datasets of the example buildings from the German residential building typology
"EnergyProfileShinyApp"  includes the datasets from the following filters: "Typology-DE_Example-Buildings", "MOBASY-All"
"All" (default):       All datasets listed above}
}
\value{
BuildingDataTables a list of dataframes including the calculation input data
"Data_Input", an empty dataframe "Data_Output" providing the structure for the output,
the dataframe "Data_Output_PreCalculated" providing data calculated by the Excel tool
(useful for comparison by developers) and the dataframe "Data_Calc" which is used to
collect all variables and their values used in the different calculation functions.
Furthermore the dataframe with all available filter lists DF_FilterBuildingData and the
actually applied filter myFilterName are returned
}
\description{
GetBuildingData_RDataPackage () loads the MOBASY building data table from the data package
"MobasyBuildingData".
GetBuildingData_RDataPackage () is a wrapper for the function GetDataMobasy () in which
arguments are used specifying the data source (local Excel or R data package)
and the building datasets to be loaded.
}
\examples{

## Load data
# Different options of dataset selection:

# (1) Load all available datasets from the MOBASY building data table
myBuildingDataTables <- GetBuildingData_RDataPackage ()
or
myBuildingDataTables <- GetBuildingData_RDataPackage ("All")

# (2) Load some example datasets
myBuildingDataTables <- GetBuildingData_RDataPackage ("Examples")

# (3) Load all datasets from the MOBASY sample
myBuildingDataTables <- GetBuildingData_RDataPackage ("MOBASY-Sample")

# (4) Load a selection of datasets from the MOBASY sample (12 buildings)
myBuildingDataTables <- GetBuildingData_RDataPackage ("MOBASY-Sample_Checked-OnSite")

# (5) Load datasets of 6 buildings from a parameter study on uncertainties
myBuildingDataTables <- GetBuildingData_RDataPackage ("ParameterStudy-CESB-2022")

# (6) Load dataset of the target/actual comparison study performed
#     for the two "PassivHausSozialPlus" (PHSP) buildings (2 x 16 variants)
myBuildingDataTables <- GetBuildingData_RDataPackage ("ParameterStudy-PHSP-2023")

# (7) Load dataset of 1 building (example for webtool)
myBuildingDataTables <- GetBuildingData_RDataPackage ("WebTool")

# (8) Load datasets of the example buildings from the German residential building typology
myBuildingDataTables <- GetBuildingData_RDataPackage ("Typology-DE_Example-Buildings")



## Show structure and content of the main dataframes
str (myBuildingDataTables$Data_Input)
str (myBuildingDataTables$Data_Output)

## Show the names of the predefined filter lists
colnames (myBuildingDataTables$DF_FilterBuildingData)

}
