% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_AuxDataTransfer.R
\name{GetDataMobasy}
\alias{GetDataMobasy}
\title{Get MOBASY data from specific sources}
\usage{
GetDataMobasy(
  Indicator_Load_ParameterTables_Excel = 1,
  Indicator_Load_ParameterTables_RDataPackage = 0,
  Indicator_Load_StationClimateTables_Excel = 1,
  Indicator_Load_StationClimateTables_RDataPackage = 0,
  Indicator_Load_BuildingData_Excel = 1,
  Indicator_Load_BuildingData_RDataPackage = 0,
  FilterVariableName_1 = "Status_DataBase_Admin",
  FilterValueList_1 = c("Public", "Project", "Test_Calc", "Analysis"),
  FilterVariableName_2 = NA,
  FilterValueList_2 = NA
)
}
\arguments{
\item{Indicator_Load_ParameterTables_Excel}{an integer (possible values: 0, 1)}

\item{Indicator_Load_ParameterTables_RDataPackage}{an integer (possible values: 0, 1)}

\item{Indicator_Load_StationClimateTables_Excel}{an integer (possible values: 0, 1)}

\item{Indicator_Load_StationClimateTables_RDataPackage}{an integer (possible values: 0, 1)}

\item{Indicator_Load_BuildingData_Excel}{an integer (possible values: 0, 1)}

\item{Indicator_Load_BuildingData_RDataPackage}{an integer (possible values: 0, 1)}

\item{FilterVariableName_1}{a character string}

\item{FilterValueList_1}{a character string or a list of character strings}

\item{FilterVariableName_2}{a character string}

\item{FilterValueList_2}{a character string or a list of character strings}
}
\value{
MobasyData a list of dataframes used for the calculation.
The following dataframes are included:
Lib_TABULA, Lib_ClimateStation, Data_Input, Data_Output_PreCalculated, Data_Output, Data_Calc
}
\description{
GetDataMobasy () loads the data used for the functions EnergyProfileCalc ()
and MobasyCalc () from sources that are specified by the parameters
(local Excel workbooks or R data packages).
}
\examples{

mydata <- GetDataMobasy (
 Indicator_Load_ParameterTables_Excel              = 1,
 Indicator_Load_ParameterTables_RDataPackage       = 0,
 Indicator_Load_StationClimateTables_Excel         = 1,
 Indicator_Load_StationClimateTables_RDataPackage  = 0,
 Indicator_Load_BuildingData_Excel                 = 1,
 Indicator_Load_BuildingData_RDataPackage          = 0,
 myFilterVariableName_1  = "ID_Dataset",
 myFilterValueList_1     = c (
   "DE.MOBASY.WBG.0008.61",
   "DE.MOBASY.WBG.0008.62",
   "DE.MOBASY.WBG.0008.63",
   "DE.MOBASY.WBG.0008.64",
   "DE.MOBASY.WBG.0008.65",
   "DE.MOBASY.WBG.0008.66"
   ),
   myFilterVariableName_2  = NA,
   myFilterValueList_2     = NA
 )

# Show structure and content of exemplary dataframes (only first row)
str (mydata$Data_Input)
str (mydata$Data_Output_PreCalculated)
str (mydata$Lib_TABULA$ParTab_InsulationDefault)
str (mydata$Lib_TABULA$ParTab_EnvArEst)
str (mydata$Lib_TABULA$ParTab_Meter_EnergyDensity)
str (mydata$Lib_ClimateStation$ClimateData_PostCodes)
str (mydata$Data_Output)

}
