% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_AuxDataTransfer.R
\name{GetParameterTables_RDataPackage}
\alias{GetParameterTables_RDataPackage}
\title{Get TABULA tables with calculation parameters from R data package}
\usage{
GetParameterTables_RDataPackage()
}
\arguments{
\item{None}{}
}
\value{
A List of 28 dataframes
}
\description{
GetParameterTables_RDataPackage () loads a data library with parameters
for the energy performance calculation from the R data package "tabuladata".
The data tables of the package are basically identical to the tables of the workbook
"tabula-values.xlsx" from the European projects TABULA and EPISCOPE
(Download at: https://episcope.eu/communication/download/)
However some tables have been added used by the MOBASY algorithms for
target/actual comparison and for uncertainty assessment.
In some of the existing tables additional parameter sets (rows) have been added.
}
\details{
GetParameterTables_RDataPackage () is a wrapper for the function Load_ParameterTables ()
in which arguments are used specifying the data source (local Excel or R data package)
}
\examples{

## Load data

TabulaTables <-
  GetParameterTables_RDataPackage ()

## Show structure and content of some of the dataframes
str (TabulaTables$ParTab_InsulationDefault)
str (TabulaTables$ParTab_System_HG)
str (TabulaTables$ParTab_Uncertainty)

}
