% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_AuxDataTransfer.R
\name{GetStationClimate_LocalExcel}
\alias{GetStationClimate_LocalExcel}
\title{Get station climate data from local Excel file}
\usage{
GetStationClimate_LocalExcel()
}
\arguments{
\item{None}{}
}
\value{
StationClimateTables a list of dataframes
}
\description{
GetStationClimate_LocalExcel () loads all station climate data from a local Excel file.
The source is fixed to:
Subfolder:    "Input/Climate"
File name:    "Gradtagzahlen-Deutschland.xlsx"
It is a wrapper for the function Load_StationClimateTables_Excel () in which
arguments are used specifying the type of data source (local Excel or R data package)
and the localisation of the local file.
}
\examples{

StationClimateTables <- GetStationClimate_LocalExcel ()

## Show structure and content of exemplary dataframes included in the list

str (StationClimateTables$ClimateData_PostCodes)
str (StationClimateTables$ClimateData_TA_HD)


}
