% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_AuxDataTransfer.R
\name{GetStationClimate_RDataPackage}
\alias{GetStationClimate_RDataPackage}
\title{Load tables with station climate data from package "clidamonger"}
\usage{
GetStationClimate_RDataPackage()
}
\arguments{
\item{None}{}
}
\value{
StationClimateTables a list of dataframes
}
\description{
GetStationClimate_RDataPackage () loads all station climate data from
the R data package clidamonger.
}
\examples{

StationClimateTables <- GetStationClimate_RDataPackage ()

## Show structure and content of exemplary dataframes included in the list

str (StationClimateTables$ClimateData_PostCodes)
str (StationClimateTables$ClimateData_TA_HD)


}
