% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_AuxDataTransfer.R
\name{\%xl_JoinStrings\%}
\alias{\%xl_JoinStrings\%}
\title{Concatenate vectors of character strings by row (Excel equivalent operator &)}
\usage{
myStr1 \%xl_JoinStrings\% myStr2
}
\arguments{
\item{myStr1}{first vector of character strings.}

\item{myStr2}{second vector of character strings.}
}
\value{
A character vector of the concatenated values.
}
\description{
%xl_JoinStrings% concatenates vectors of character strings by row.
The function is used to simplify parsing Excel formulas and to apply them to vector variables.
Definition of infix operators: https://www.datamentor.io/r-programming/infix-operator/
}
\examples{
temp1 <- c ("cow ", "pig ", "elephant ", "tiger ")
temp2 <- c ("eats ", "loves ", "grows ", "has ")
temp3 <- c ("grass", "mud", "big", "stripes")
temp1 \%xl_JoinStrings\% temp2 \%xl_JoinStrings\% temp3
# Result: "cow eats grass"     "pig loves mud"      "elephant grows big" "tiger has has "

}
