% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Main.R
\name{MobasyCalc}
\alias{MobasyCalc}
\title{Perform full MOBASY energy performance calculation  (including local climate, uncertainties and target/actual comparison)}
\usage{
MobasyCalc(TabulaTables, myBuildingDataTables, StationClimateTables = NA)
}
\arguments{
\item{TabulaTables}{a list of dataframes with parameters used for the calculation}

\item{myBuildingDataTables}{a list of dataframes including the calculation input data
"Data_Input", an empty dataframe "Data_Output" providing the structure for the output,
the dataframe "Data_Output_PreCalculated" providing data calculated by the Excel tool
(useful for comparison by developers) and the dataframe "Data_Calc" which is used to
collect all variables and their values used in the different calculation functions.}

\item{StationClimateTables}{a list of dataframes with climate data from local weather stations}
}
\value{
myOutputTables a list of two output dataframes Data_Output (main results in
a predefined structure) and Data_Calc (values of all temporary variables)
}
\description{
MobasyCalc consists of a physical model for calculating the energy demand for heating and domestic hot water.
Energy Profile Indicators or detailed building data may be used as input data.
The energy performance is calculated by use of German local climate by postcode.
The calculation also includes an uncertainty assessment
and a comparison with measured consumption data (target/actual comparison).
(If a simplified version is needed the function EnergyProfileCalc () can be used, which is a subset of MobasyCalc ()
}
\details{
An overview of the method can be found in:
Loga, Tobias; Stein, Britta; Behem, Guillaume: Use of Energy Profile Indicators to Determine
the Expected Range of Heating Energy Consumption; Proceedings of the Conference
"Central Europe towards Sustainable Building" 2022 (CESB22),
4 to 6 July 2022; Acta Polytechnica CTU Proceedings 38:470–477, 2022
https://ojs.cvut.cz/ojs/index.php/APP/article/view/8299/6839
https://doi.org/10.14311/APP.2022.38.0470
}
\examples{

## Get local parameter tables
TabulaTables <- GetParameterTables_LocalExcel ()


## Get data from weather stations
StationClimateTables <- GetStationClimate_LocalExcel ()


## Get local building data
# Different options of dataset selection:

# (1) Load all available datasets from the MOBASY building data table
myBuildingDataTables <- GetBuildingData_LocalExcel ()

# (2) Load all datasets from the MOBASY sample
myBuildingDataTables <- GetBuildingData_LocalExcel ("MOBASY-Sample")

# (3) Load dataset of 1 building (example for webtool)
myBuildingDataTables <- GetBuildingData_LocalExcel ("WebTool")

# (4) Load dataset of the target/actual comparison study performed
#     for the two "PassivHausSozialPlus" (PHSP) buildings (2 x 16 variants)
myBuildingDataTables <- GetBuildingData_LocalExcel ("ParameterStudy-PHSP-2023")

# (5) Load datasets of 6 buildings from a parameter study on uncertainties
myBuildingDataTables <- GetBuildingData_LocalExcel ("ParameterStudy-CESB-2022")


## Calculate energy performance (MOBASY calculation)

myOutputTables <- MobasyCalc (
   TabulaTables,
   myBuildingDataTables,
   StationClimateTables
   )


## Show structure and content of the two output dataframes

str (myOutputTables$Data_Output)
str (myOutputTables$Data_Calc)

}
