% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_AuxDataTransfer.R
\name{Save_BuildingData_LocalExcel}
\alias{Save_BuildingData_LocalExcel}
\title{Save the building data tables to an Excel workbook}
\usage{
Save_BuildingData_LocalExcel(Data_Output, Data_Calc, Data_Output_PreCalculated)
}
\arguments{
\item{myBuildingDataTables}{a list of data frames. THese are:
Data_Input, Data_Output_PreCalculated, Data_Output, Data_Calc}
}
\description{
Save all current building data including calculation results to an Excel workbook
}
\details{
The following tables are exported to the respective subfolders:

Subfolder "Output/Calc/":

> "Data_Calc_{YYYY-MM-DD_hh-mm-ss}.xlsx"


Subfolder "Output/Result/":

> "Data_Input_{YYYY-MM-DD_hh-mm-ss}.xlsx"

> "Data_Output_{YYYY-MM-DD_hh-mm-ss}.xlsx"

> "Data_Output_PreCalculated_{YYYY-MM-DD_hh-mm-ss}.xlsx"
}
\examples{

## Load building data tables and calculation parameters

## Get parameter tables
TabulaTables <-
   GetParameterTables_RDataPackage ()

myBuildingDataTables <-
   GetBuildingData_RDataPackage ("Examples")

myOutputTables <- EnergyProfileCalc (
   TabulaTables,
   myBuildingDataTables
   )

Save_BuildingData_LocalExcel (
   myOutputTables$Data_Output,
   myOutputTables$Data_Calc,
   myBuildingDataTables$Data_Output_PreCalculated
)

}
