% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_AuxDataTransfer.R
\name{Save_BuildingData_rda}
\alias{Save_BuildingData_rda}
\title{Save building data as .rda in a subfolder of the working directory}
\usage{
Save_BuildingData_rda(
  Data_Input,
  Data_Output_PreCalculated,
  Data_Output,
  Header_BuildingData,
  DF_FilterBuildingData,
  myFilterName
)
}
\arguments{
\item{myBuildingDataTables}{a list of data frames. It includes:
Data_Input, Data_Output_PreCalculated, Data_Output, Data_Calc}
}
\description{
Save all current building data including calculation results as *.rda files
}
\details{
The following tables are exported to the respective subfolders:
"Data_Input":                "Output/RDA/"
"Data_Output":               "Output/RDA/"
"Data_Output_PreCalculated": "Output/RDA/"
"Data_Calc":                 "Output/RDA/"
}
\examples{

## Load building data tables and calculation parameters

## Get parameter tables
TabulaTables <-
   GetParameterTables_RDataPackage ()

myBuildingDataTables <-
   GetBuildingData_RDataPackage ("Examples")

myOutputTables <- EnergyProfileCalc (
   TabulaTables,
   myBuildingDataTables
   )

Save_BuildingData_rda (
   myBuildingDataTables$Data_Input,
   myBuildingDataTables$Data_Output_PreCalculated,
   myOutputTables$Data_Output,
   myBuildingDataTables$Header_BuildingData,
   myBuildingDataTables$DF_FilterBuildingData,
   myBuildingDataTables$myFilterName
)

}
