% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_AuxDataTransfer.R
\name{Save_BuildingData_LocalExcel}
\alias{Save_BuildingData_LocalExcel}
\title{Save the building data tables to an Excel workbook}
\usage{
Save_BuildingData_LocalExcel(Data_Output, Data_Calc, Data_Output_PreCalculated)
}
\arguments{
\item{myBuildingDataTables}{a list of data frames. THese are:
Data_Input, Data_Output_PreCalculated, Data_Output, Data_Calc}
}
\description{
Save all current building data including calculation results to an Excel workbook
}
\details{
The following tables are exported to the respective subfolders:
"Data_Input":                "Output/Result/"
"Data_Output":               "Output/Result/"
"Data_Output_PreCalculated": "Output/Result/"
"Data_Calc":                 "Output/Calc/"
}
\examples{

## Load building data tables and calculation parameters

## Get parameter tables
TabulaTables <-
   GetParameterTables_RDataPackage ()

myBuildingDataTables <-
   GetBuildingData_RDataPackage ("Examples")

myOutputTables <- EnergyProfileCalc (
   TabulaTables,
   myBuildingDataTables
   )

Save_BuildingData_LocalExcel (
   myBuildingDataTables$Data_Output,
   myBuildingDataTables$Data_Calc,
   myBuildingDataTables$Data_Output_PreCalculated
)

}
