% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_AuxDataTransfer.R
\name{Save_BuildingData_rda}
\alias{Save_BuildingData_rda}
\title{Save building data as .rda in a subfolder of the working directory}
\usage{
Save_BuildingData_rda(
  Data_Input,
  Data_Output_PreCalculated,
  Data_Output,
  Header_BuildingData,
  DF_FilterBuildingData,
  myFilterName,
  SubFolderName = "Output/RDA"
)
}
\arguments{
\item{myBuildingDataTables}{a list of data frames. It includes:
Data_Input, Data_Output_PreCalculated, Data_Output, Data_Calc}
}
\description{
Save all current building data including calculation results as *.rda files
}
\details{
The following tables are exported to the subfolder "Output/RDA/":

    Data_Input.rda

    Data_Output_PreCalculated.rda

    Data_Output.rda

    DF_FilterBuildingData.rda

    myFilterName.rda

    InfoVariables.rda
}
\examples{

## Example 1:
## Load parameters and building data, calculate and save input and output to RDA files

TabulaTables <-
   GetParameterTables_RDataPackage ()

myBuildingDataTables <-
   GetBuildingData_RDataPackage ("Examples")

myOutputTables <- EnergyProfileCalc (
   TabulaTables,
   myBuildingDataTables
   )

Save_BuildingData_rda (
   myBuildingDataTables$Data_Input,
   myBuildingDataTables$Data_Output_PreCalculated,
   myOutputTables$Data_Output,
   myBuildingDataTables$Header_BuildingData,
   myBuildingDataTables$DF_FilterBuildingData,
   myBuildingDataTables$myFilterName
)


## Example 2:
## Load local building data from Excel and save as RDA files used for input of calculation

# The data frame "Data_Output" is empty and is used frame for the calculation output

myBuildingDataTables <-
   GetBuildingData_LocalExcel ("All")

Save_BuildingData_rda (
   myBuildingDataTables$Data_Input,
   myBuildingDataTables$Data_Output_PreCalculated,
   myBuildingDataTables$Data_Output,
   myBuildingDataTables$Header_BuildingData,
   myBuildingDataTables$DF_FilterBuildingData,
   myBuildingDataTables$myFilterName
)

}
